+++ src/library/math/operator/exports.ts
@@ -1,5 +1,5 @@
 export * from './predicate/exports';
 
-export * from './hyper';
+export * from './hyper/exports';
 
 export * from './constructive/exports';
