+++ src/library/math/operator/hyper/exports.ts
